add: constants move: prices and controls to constants
