+++ src/components/Burger/BuildControls/BuildControls.js
@@ -1,19 +1,20 @@
 import React from 'react';
 import styles from './BuildControls.module.css';
 import BuildControl from './BuildControl/BuildControl';
+import CONTROLS from '../../../constants/controls';
 
 //ингредиенты для списка
-const controls = [
-  {label: 'Salad', type: 'salad'},
-  {label: 'Meat', type: 'meat'},
-  {label: 'Bacon', type: 'bacon'},
-  {label: 'Cheese', type: 'cheese'},
-];
+// const controls = [
+//   {label: 'Salad', type: 'salad'},
+//   {label: 'Meat', type: 'meat'},
+//   {label: 'Bacon', type: 'bacon'},
+//   {label: 'Cheese', type: 'cheese'},
+// ];
 
 const BuildControls = (props) => {
   return (
     <div className={styles.BuildControls}>
-      {controls.map(({label, type}) => (
+      {CONTROLS.map(({label, type}) => (
         <BuildControl
           key={label}
           label={label}
