+++ src/constants/controls.js
@@ -1,0 +1,8 @@
+const CONTROLS = [
+  {label: 'Salad', type: 'salad'},
+  {label: 'Meat', type: 'meat'},
+  {label: 'Bacon', type: 'bacon'},
+  {label: 'Cheese', type: 'cheese'},
+];
+
+export default CONTROLS;
