+++ src/constants/ingridient-prices.js
@@ -1,0 +1,8 @@
+const INGREDIENT_PRICES = {
+  salad: 0.5,
+  bacon: 0.7,
+  meat: 1.5,
+  cheese: 0.5,
+};
+
+export default INGREDIENT_PRICES;
